+++ Web/js/app/controllers/technician_manager.js
@@ -14,7 +14,19 @@ $(document).ready(function() {
       if (key === "edit") {
         technician_manager.retrieveTechnician(options.$trigger);
       } else if (key === "delete") {
-        technician_manager.delete(parseInt(options.$trigger.attr("data-technician-id")));
+				var msg = $('#confirmmsg-delete').val();
+				if (typeof msg !== typeof undefined && msg !== false) {
+					utils.showConfirmBox(msg, function(result){
+						if(result)
+						{
+							technician_manager.delete(parseInt(options.$trigger.attr("data-technician-id")));
+						}
+					});
+				}
+				else
+				{
+					technician_manager.delete(parseInt(options.$trigger.attr("data-technician-id")));
+				}
       }
     },
     items: {
@@ -44,7 +56,20 @@ $(document).ready(function() {
     }
   });
   $(".from-inactive-list").click(function() {
-    technician_manager.updateTechnicians("active", technician_ids);
+		var msg = $('#confirmmsg-activate').val();
+		if (typeof msg !== typeof undefined && msg !== false) {
+			utils.showConfirmBox(msg, function(result){
+				if(result)
+				{
+					technician_manager.updateTechnicians("active", technician_ids);
+				}
+			});
+		}
+		else
+		{
+			technician_manager.updateTechnicians("active", technician_ids);
+		}
+    
   });
   $(".from-active-list").click(function() {
     technician_manager.updateTechnicians("inactive", technician_ids);
@@ -76,7 +101,19 @@ $(document).ready(function() {
   });//Edit a technician
 
   $("#btn_delete_technician").click(function() {
-    technician_manager.delete(parseInt(utils.getQueryVariable("technician_id")));
+		var msg = $('#confirmmsg-delete').val();
+		if (typeof msg !== typeof undefined && msg !== false) {
+			utils.showConfirmBox(msg, function(result){
+				if(result)
+				{
+					technician_manager.delete(parseInt(utils.getQueryVariable("technician_id")));
+				}
+			});
+		}
+		else
+		{
+			technician_manager.delete(parseInt(utils.getQueryVariable("technician_id")));
+		}
   });//Delete a technician
 
   if (utils.getQueryVariable("mode") === "edit") {
@@ -103,11 +140,6 @@ $(document).ready(function() {
     $(".technician_list").fadeIn('2000').removeClass("hide");
     technician_manager.getList();
   });//Show "List All" panel
-  
-  $("li[data-technician-id]").hover(function(){
-	$(this).tooltip({ placement: "bottom" });
-  }); 
-
 });
 /***********
  * technician_manager namespace
